requirements.txt
appdirs==1.4.0
Django==1.10.5
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
python-decouple==3.0
six==1.10.0
whitenoise==3.3.0
